scholar_to_bibtex/requirements.txt
google-search-results==2.4.2
Flask
python-dotenv